routinepush

+++ montecpp/include/outputFunction.hpp
@@ -12,16 +12,8 @@ class OutputFunction : public SimulatorFunction {
                    const std::function<double(const std::vector<double>&)> _outputFunction)
         : inputNames{_inputNames}, numInputs{_inputNames.size()}, SimulatorFunction(_name, _outputFunction){};
 
-    //  using SimulatorFunction::attachInputs;
-    //  using SimulatorFunction::compute;
-    //  using SimulatorFunction::data;
-    //  using SimulatorFunction::name;
-    //  using SimulatorFunction::samples;
-    //  using SimulatorFunction::operator[];
-    //  using SimulatorFunction::getData;
-
     std::vector<std::string> inputNames;
-    const int numInputs;
+    const size_t numInputs;
 
    private:
 };
